UCT/IB/MLX5/DC: config dcis_init_capacity should be unsigned long to support auto

+++ src/uct/ib/mlx5/dc/dc_mlx5.h
@@ -192,7 +192,7 @@ typedef struct uct_dc_mlx5_iface_config {
     ucs_time_t                          fc_hard_req_timeout;
     uct_ud_mlx5_iface_common_config_t   mlx5_ud;
     unsigned                            num_dci_channels;
-    unsigned                            dcis_initial_capacity;
+    unsigned long                       dcis_initial_capacity;
 } uct_dc_mlx5_iface_config_t;
